Drop support for MariaDB 10.2 and 10.3
